+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [1.10.0](https://github.com/hey-car/heycar-uikit/compare/v1.9.0...v1.10.0) (2022-08-03)
+
+
+### Features
+
+* **collapse:** the first version of the collapse component ([#46](https://github.com/hey-car/heycar-uikit/issues/46)) ([4d10244](https://github.com/hey-car/heycar-uikit/commit/4d102446b2c6348408bb3bd1d5a7ebf1ec1cd6f5))
+
+# [1.9.0](https://github.com/hey-car/heycar-uikit/compare/v1.8.0...v1.9.0) (2022-08-03)
+
+
+### Features
+
+* **badge:** The first implantation of badge component ([#42](https://github.com/hey-car/heycar-uikit/issues/42)) ([2ca5a7e](https://github.com/hey-car/heycar-uikit/commit/2ca5a7ea897c4cdeccaadca4f4d1592be40fe6e7))
+
+# [1.8.0](https://github.com/hey-car/heycar-uikit/compare/v1.7.0...v1.8.0) (2022-08-02)
+
+
+### Features
+
+* **icons:** updated icons ([#45](https://github.com/hey-car/heycar-uikit/issues/45)) ([2293b7a](https://github.com/hey-car/heycar-uikit/commit/2293b7ad0135fb645495afb9a11583ce8ef7e6e3))
+
 # [1.7.0](https://github.com/hey-car/heycar-uikit/compare/v1.6.0...v1.7.0) (2022-07-27)